pyproject.toml
[project]
name = "kero-voice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch==2.7.0",
    "torchvision==0.22.0",
    "torchaudio==2.7.0",
    "python-dotenv>=1.2.1",
    "discord-py>=2.6.4",
    "tts>=0.22.0",
    "pynacl>=1.6.1",
    "cutlet>=0.5.0",
    "fugashi>=1.5.2",
    "unidic-lite>=1.0.8",
    "transformers==4.36.2",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu128" }]
torchvision = [{ index = "pytorch-cu128" }]
torchaudio = [{ index = "pytorch-cu128" }]